Refactor: BaggageBundles wrapper
Summary: 
- Removed Baggage Card wrapper (it is not necessary according to current designs).
- Removed horizontal margin to align items edges with inputs.
- Updated `BaggageBundle` to make whole component clickable (only bullet can be clicked in previous version).
