+++ apps/core/scenes/passengerForm/baggageBundles/BaggageTypeDetail.js
